Prevent drag scolling in Firefox
When dragging an element, dragging past the end of it's container causes
the container to scoll if has an overflow set.
This is supressed in Chrome as chrome uses mousemove events to perform
the dragging.
Firefox seems to do some magic with mousedown on the otherhand.

This makes the behaviour consistent and non-annoying - however it would
be nice to keep scrolling and scroll the correct tree instead.